code-for-sage/requirements.txt
# Core Framework
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0

# AI/ML
openai==1.3.0
spacy==3.7.2
scikit-learn==1.3.2
transformers==4.36.2

# Database
sqlalchemy==2.0.37
alembic==1.13.1

# Data Processing & Visualization
pandas==2.1.4
numpy==1.26.2
plotly==5.18.0

# HTTP/API
requests==2.31.0
aiohttp==3.9.1
python-jose==3.3.0

# Testing & Development
pytest==7.4.3
pytest-asyncio==0.23.2
flake8==6.1.0
black==23.12.1

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# Security
cryptography==41.0.7
python-jose[cryptography]==3.3.0

# Web Scraping Support
beautifulsoup4==4.12.2
lxml==4.9.3
